Finish creating the deck
First attempt failed.
1st, no sort field which could be problem.
2nd, the format of the filenames was wrong. it should just be the name
of the file and should be surrounded like this: [sound:filename.mp3].
This one did it.

I am now uploading the files to ankiweb and it's a lot. I don't know if
I'll overload the damn server, so far as I know there is no limit. But
this is pretty big. Scratch that, it's just 75 megs. The cards
themselves might be a couple megs as well (there's 8000 something
cards).